requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
authlib==1.7.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
black==26.3.1
blinker==1.9.0
blis==1.3.3
build==1.5.0
cachetools==7.1.1
catalogue==2.0.10
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
cloudpathlib==0.24.0
colorama==0.4.6
comm==0.2.3
confection==1.3.3
cryptography==47.0.0
cymem==2.0.13
dataclasses-json==0.6.7
ddgs==9.14.2
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring-parser==0.18.0
duckduckgo-search==8.1.1
durationpy==0.10
emoji==2.15.0
et-xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.13.2
filelock==3.29.0
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.4.0
gitdb==4.0.12
gitpython==3.1.49
google-auth==2.50.0
google-genai==1.74.0
googleapis-common-protos==1.74.0
greenlet==3.5.0
groq==0.37.1
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.3
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.13.0
idna==3.13
importlib-metadata==8.7.1
importlib-resources==7.1.0
iniconfig==2.3.0
installer==0.7.0
ipykernel==7.2.0
ipython==9.13.0
ipython-pygments-lexers==1.1.1
isort==8.0.1
itsdangerous==2.2.0
jedi==0.20.0
jinja2==3.1.6
jiter==0.14.0
joserfc==1.6.4
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
kubernetes==35.0.0
langchain==1.2.17
langchain-anthropic==1.4.3
langchain-classic==1.0.5
langchain-community==0.4.1
langchain-core==1.3.2
langchain-experimental==0.4.1
langchain-google-genai==4.2.2
langchain-groq==1.1.2
langchain-huggingface==1.2.2
langchain-openai==1.2.1
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langchain-xai==1.2.2
langdetect==1.0.9
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.13
langsmith==0.8.0
llvmlite==0.47.0
lxml==6.1.0
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
murmurhash==1.0.15
mypy-extensions==1.1.0
narwhals==2.20.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.25.1
openai==2.33.0
openpyxl==3.1.5
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.2
pandas==3.0.2
parso==0.8.7
pathspec==1.1.1
pdf2image==1.17.0
pillow==12.2.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.9.6
playwright==1.59.0
pluggy==1.6.0
preshed==3.0.13
primp==1.2.3
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pure-eval==0.2.3
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1-modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.13.3
pydantic-core==2.46.3
pydantic-settings==2.14.0
pydeck==0.9.2
pyee==13.0.1
pygments==2.20.0
pymupdf==1.27.2.3
pypdf==6.10.2
pypdfium2==5.7.1
pypika==0.51.1
pyproject-hooks==1.2.0
pytesseract==0.3.13
pytest==9.0.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2026.4.20
python-magic==0.4.27
python-multipart==0.0.27
python-oxmsg==0.0.2
pytokens==0.4.1
pyyaml==6.0.3
pyzmq==27.1.0
rapidfuzz==3.14.5
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart-open==7.6.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.49
srsly==2.5.3
stack-data==0.6.3
starlette==1.0.0
streamlit==1.57.0
tavily-python==0.7.24
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
tokenizers==0.23.1
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.25.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2026.2
unstructured==0.22.26
unstructured-client==0.43.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.46.0
validators==0.35.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.7.0
weasel==1.0.0
weaviate-client==4.21.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wikipedia==1.4.0
wrapt==2.1.2
xlrd==2.0.2
xxhash==3.7.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0